Launch $SHELL by default if it is set
In an user environment (which saneterm would usually be executed in)
$SHELL should point to the current user's login shell as per environ(7).
On startup, check if the variable is set and use it to populate the
default command to run if it is, otherwise fall back to `sh` like
before.

I believe this is the correct change even if it gives users (including
me) more trouble since many login shells people will use (like fish)
don't work well in saneterm, whereas /bin/sh usually will work fine.