Clear the line if Down is pressed at the most recent history entry
This mimicks the behavior familiar from most shells. We also make sure
that this doesn't happen when we hit the oldest history entry -- in that
case we will just do nothing if Up is pressed.

To implement this correctly we need to distinguish between a position in
the history and no position in the history. We encode the latter using
`None` and just call `reset_history_index()` in case we encounter it to
reacquire a point of reference.