Add workaround for annoying scrollbar if wordwrap is enabled
There is an issue [1] in GTK 3 which causes the automatic hyphenation of
GtkTextViews to cause a horizontal scroll bar to appear despite it being
not necessary (in our case the horizontal scrollbar is rendered over the
hyphens which is hardly cause for concern).

Ideally we would just disable auto hyphenation, but this doesn't seem to
be possible [2] in GTK 3 yet. As a workaround we just disable horizontal
scrolling altogether in in update_wrapmode() if wordwrap is enabled.

While we're at it add a note about the GTK 4 feature [3] which would
allow us to disable auto hyphenation. For GTK 3 it seems to require to
set a pango [4] attribute in the pango layout used for rendering the
widget which is not exposed in a modifiable way by GtkTextView
unfortunately -- or at least I haven't been able to figure out how the
desired result could be achieved.

[1]: https://gitlab.gnome.org/GNOME/gtk/-/issues/2384
[2]: https://gitlab.gnome.org/GNOME/gtk/-/issues/2530
[3]: https://docs.gtk.org/gtk4/property.TextTag.insert-hyphens.html
[4]: See these solutions of other projects that ran into this issue:
     https://github.com/mate-desktop/caja/pull/1334
     https://gitlab.gnome.org/GNOME/nautilus/-/merge_requests/482