termview: update _last_mark even on newline
Most REPLs will write a new prompt on newline but when starting a
graphical application from a shell (e.g. gitk) no output will be written
to the PTY after the newline. This causes _last_mark to not be updated
and causes cursor_at_end to always return False making it impossible to
quit the application using termios.VINTR.