Support bell character by parsing pty input into events
This commit introduces a new object, the PtyParser, which will be
handling pty input in the future. It receives (utf-8 decoded) input
and emits a series of events which encode pty output we want to
represent in the UI:

* The most basic event type is TEXT which is plain text we need to add
  to the TermBuffer.

* Additionally BELL is now supported which indicates we want to beep
  (if enabled in gtk) and set the urgency flag of the window.

Events seems to be a simple way to indicate future special behavior
and by splitting the data read from the pty we can easily skip control
characters which we don't want to render (and thus get rid of ugly
codepoint boxes). Due to the use of yield and string slices, this
approach should also be efficient enough.

Using events instead of triggering the desired action in handle_pty
ad-hoc also has the advantage that we can split the parsing logic into
a separate object which is interesting for the following reasons:

* Especially when we want to support a subset of ANSI control
  sequences we'll need to track parser state:

  - ANSI escape sequences are multiple codepoints long, so we need to
    track a parser state in order to keep it incremental.

  - ANSI escape sequences allow changing the font properties (style,
    weight, color) which is implemented in a stateful way, we'll also
    need to keep track of.

* The parser is implemented independently of the rest of the
  application and especially the UI, so we'll be able to unit
  test it easily.