pty.Parser: parse and filter out CSI escape sequences
Add a parser for CSI escape sequences as specified in ECMA-48 (5th
edition). All other escape sequences starting with \033 (like OSC
escape sequences) aren't supported yet and will be added to the
terminal's buffer like before.

No CSI escape sequence generates a dedicated event so far, the only
effect of them being parsed is that they are omitted from the output
in the terminal. This makes the output of programs making heavy use of
e. g. SGR escape sequences much more readable.

If the end of input is hit while parsing an escape sequence, we
postpone parsing it until the next invocation of parse().

Additionally we rely on backtracking to deal with unknown and invalid
escape sequences: By setting a hint that the sequence is to be ignored
and backtracking, the rest of the parser behaves normally and emits
the invalid/unknown sequence as normal text.