Track text style in Terminal and reuse TextTags
This change addresses the previous layer violation mentioned in a
previous commit: Instead of inflating the Parser code by tracking
state which is not necessary for parsing, we now just emit events
describing text style changes in the parser code. The necessary
updates to actual display state is then handled in the Terminal
object.

We use this to map the changes to display style introduced by events
to _individual_ TextTags instead of creating a composite TextTag for
every styled stretch of text.

This makes TextTags reusable since they are less specific: Instead of
a single italic, green background and bold face tag, we now have three
different ones which are quite likely to occur again. Thus we'll store
(references to) already created tags in a dictionary to be reused.

Note that we can't use Gtk's TextTag lookup mechanism which allows to
reuse TextTags in the TextTagTable of a buffer by name since this
would not allow us to create TextTags on demand. Precreating all 13824
possible color tags is of course not an option.

While this change doesn't bring a big performance improvement it
should prevent performance from degrading over time since it limits
the number of tags being created, allocated and managed in the Gtk
TextBuffer.