Move parsing of CSI escape sequences into dedicated functions
This breaks up the deeply indented blocks which implemented this in
pty.Parser.parse() before. It also allows us to simplify the control
flow a bit: The two nested try … except blocks for StopIteration (to
detect if we need to postpone parsing) and AssertionError (to detect
if we don't understand what looked like an escape sequence or handle
other parse errors) into a single one.