Generate max. one event for a SGR sequence with colons as seperators
Sequences with colons are used when only the first parameter
determines the kind of SGR event caused by the sequence and the rest
of the parameters are additional arguments. This means that the
arguments can contain numbers which are meaningful by themselves, like
0 (reset).

This means that if such a sequence is not supported by saneterm we
previously erroneously interpreted arguments in such a way that we
generated wrong events. This happened for example in vte's sgr-test.sh
which relied on kitty's non standard underline escape sequences which
we don't support at this time:
https://sw.kovidgoyal.net/kitty/protocol-extensions.html#colored-and-styled-underlines